# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "version": "0.1"
    },
    "paths": {
        "/api/pedido": {
            "get": {
                "tags": [
                    "pedido"
                ],
                "operationId": "1c643b61b31c4b066aa2f723b2c755bc",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "pedido"
                ],
                "summary": "Criar venda",
                "operationId": "9e3bd6b7e6c3fed6d7fa7d0720ac3220",
                "parameters": [
                    {
                        "name": "vendedor_id",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    },
                    {
                        "name": "valor",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/pedido/{id}": {
            "get": {
                "tags": [
                    "pedido"
                ],
                "operationId": "a1d998e3688f711975a47999b5a70c8f",
                "responses": {
                    "200": {
                        "description": "Mostrar Pedido"
                    }
                }
            },
            "put": {
                "tags": [
                    "pedido"
                ],
                "operationId": "03ff808340e6a4bb1e2ce007abcff75a",
                "responses": {
                    "200": {
                        "description": "Atualizar Pedido"
                    }
                }
            },
            "delete": {
                "tags": [
                    "pedido"
                ],
                "operationId": "44b8c42bfba9cde9749d9822ddb68262",
                "responses": {
                    "200": {
                        "description": "Excluir Pedido"
                    }
                }
            }
        },
        "/api/pedidos/totalDia": {
            "get": {
                "tags": [
                    "pedido"
                ],
                "summary": "Mostrar o total de vendas do dia",
                "operationId": "62c82d4c0165621e0af160dc62bc0eb9",
                "parameters": [
                    {
                        "name": "data",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "13-06-2023"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/produto": {
            "get": {
                "tags": [
                    "produto"
                ],
                "operationId": "9e8b6153520e2c11e39c18d32a728581",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "produto"
                ],
                "operationId": "6ca322ebf4a8ee6f7f9c74a3a5a19cf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "estoque",
                                    "preco"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Caneta azul"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Caneta azul ponta fina"
                                    },
                                    "estoque": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "100"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/produto/{id}": {
            "get": {
                "tags": [
                    "produto"
                ],
                "operationId": "1e02935046408bfa9cde9ce7d2b68805",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "produto"
                ],
                "operationId": "79676372602f656d77d498d2fe5aa44b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "estoque",
                                    "preco"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Caneta azul"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Caneta azul ponta fina"
                                    },
                                    "estoque": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "100"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "produto"
                ],
                "operationId": "216f91a671205269378e26fb3310044b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do produto",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/vendedor": {
            "get": {
                "tags": [
                    "vendedor"
                ],
                "operationId": "188c7f38c253ecf071cf315d7739bc7f",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "vendedor"
                ],
                "operationId": "689c564a1a842204285757211e299bdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@hotmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/vendedor/{id}": {
            "get": {
                "tags": [
                    "vendedor"
                ],
                "operationId": "2b999353b6b2004f6fae3b9e3aea01f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do vendedor",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "vendedor"
                ],
                "operationId": "e6753b96ab4de0923ff6d2ea8bfd3b5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do vendedor",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@hotmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "vendedor"
                ],
                "operationId": "30c8774ec7fc6eed75d9dbece751e4c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do vendedor",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/vendedor/pedidos/comissao": {
            "get": {
                "tags": [
                    "vendedor"
                ],
                "summary": "Listar todos os vendedores",
                "operationId": "329eb4de07b3ab412248d4533520d95d",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/api/vendedor/{id}/pedidos": {
            "get": {
                "tags": [
                    "vendedor"
                ],
                "summary": "Mostrar todas as vendas do vendedor",
                "operationId": "698c651f1d2f018f37efbb2b8dda4dc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do vendedor",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        }
    }
}